Implements a data mode - can enter data manually - copy/paste to itself and to from spreadsheet editors

+++ Desktop/analysis/analysis.cpp
@@ -172,6 +172,9 @@ void Analysis::run()
 
 void Analysis::refresh()
 {
+	if(_refreshBlocked)
+		return;
+
 	clearRSources();
 	TempFiles::deleteAll(_id);
 	run();
